requirements.txt
streamlit==1.35.0
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
seaborn==0.13.2
scikit-learn==1.4.2
joblib==1.4.2
fpdf==1.7.2
Pillow==10.3.0
pdf2image==1.17.0
easyocr==1.7.1
torch==2.3.0
opencv-python==4.9.0.80